# platform/data_api/routes/candles.py
# -------------------------------------------------------------------------------------------------
#  Copyright (C) 2015-2026 Nautech Systems Pty Ltd. All rights reserved.
#  https://nautechsystems.io
#
#  Licensed under the GNU Lesser General Public License Version 3.0 (the "License");
#  You may not use this file except in compliance with the License.
#  You may obtain a copy of the License at https://www.gnu.org/licenses/lgpl-3.0.en.html
#
#  Unless required by applicable law or agreed to in writing, software
#  distributed under the License is distributed on an "AS IS" BASIS,
#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
#  See the License for the specific language governing permissions and
#  limitations under the License.
# -------------------------------------------------------------------------------------------------
"""
`GET /api/candles/{instrument_id}` -- cursor-paginated candle history (AD-F3): the
frontend's chart page fetches its initial 120-bar window and every scroll-back page
through this one `before_ns`/`limit` contract, never an unbounded full-range load.

Format + transport only (Story 24.2): the page itself is `views.chart_series.candle_page` (the
candle store first, the archive's one seconds -> bars fold for what it does not cover) and its
gap rows `views.chart_series.with_gap_markers`. This module clamps the query params, passes its
own `CATALOG_PATH`/`CANDLES_DB_DIR` and the live bus's unflushed tail in, builds the response
models, and maps `ImpossibleCandle` to a 500. A catalog file the read cannot use (truncated or
corrupt, or files that kept disappearing across its listings: `kernel.catalog_files.CatalogReadError`)
is ledgered at `data_api.candles_catalog_read` and a 500 too (DATA-07): the chart shows an error,
never a page missing that file's seconds.

`CATALOG_PATH` comes from `data_api.settings` (a leaf module -- routes can't import it from
`app.py`, which imports them).
"""

from fastapi import APIRouter
from fastapi import HTTPException
from kernel import catalog_files
from kernel.venues import market_kind
from kernel.venues import venue_of
from observability import error_ledger
from pydantic import BaseModel
from views import chart_series
from views.catalog_reads import NoInstrumentDefinition
from views.catalog_reads import instrument_precision

from data_api import buses
from data_api.settings import CANDLES_DB_DIR
from data_api.settings import CATALOG_PATH


# Server-enforced upper bound on `limit`, regardless of what the client requests (AC #7,
# MEM-01 extended to the API surface) -- a module constant, not per-request configurable.
_MAX_CANDLES_LIMIT = 500

# Server-enforced bound on `bar_seconds` -- same silent-clamp philosophy as `limit` above
# (AC #7's "no error, silently clamped"), not a 422: a non-positive value would zero or
# invert every window/bucket computation, and an unbounded one would let a client
# blow up the query span this route otherwise keeps deliberately bounded (AD-F3/MEM-01).
_MAX_BAR_SECONDS = 604_800  # 1w -- the timeframe selector's widest bar

router = APIRouter()


class CandleItem(BaseModel):
    t: int
    o: float | None = None
    h: float | None = None
    l: float | None = None
    c: float | None = None
    v: float | None = None
    # Bucket observed for < 90% of its span (collector gaps, D-15): its high/low/volume are
    # understated. Set on every candle, stored or folded from raw 1s at read time (Story 31.8);
    # absent (None) on gap markers only.
    partial: bool | None = None
    # Story 33.3 (`docs/DATA_DICTIONARY.md` §2.15): the bar's exact order-flow and liquidation sums,
    # integers in units of this item's own `price_precision`/`size_precision` (the bar's finest,
    # which may differ from the response's definition precisions after a venue changed them):
    # `buy_v`/`sell_v`/`liq_long_v`/`liq_short_v` count `10^-size_precision`, `pv` (the
    # second-close VWAP numerator) `10^-(price_precision + size_precision)`, the `_n` keys trades
    # and liquidations. Null is unknown, never 0: a pre-migration bar's flow, an instrument without
    # a liquidation feed's `liq_*` (Bybit linear only has one), a feed instrument's `liq_*` before
    # its first archived liquidation (audit D-160), and every key of a gap marker.
    # Known limit (JSON range): `pv` (and a very large volume) can exceed 2^53, which a browser's
    # `JSON.parse` rounds to the nearest double; Python consumers (`verification`, research) get
    # the exact int. The UI uses `pv` only as a VWAP numerator, `pv / (buy_v + sell_v)`, where a
    # 2^-53 relative error is invisible. Upgrade path: a string encoding of the integer columns
    # (an added key per AD-D12), parsed with `BigInt` where exactness matters in the browser.
    buy_v: int | None = None
    sell_v: int | None = None
    buy_n: int | None = None
    sell_n: int | None = None
    pv: int | None = None
    liq_long_v: int | None = None
    liq_short_v: int | None = None
    liq_n: int | None = None
    price_precision: int | None = None
    size_precision: int | None = None


class CandlesResponse(BaseModel):
    items: list[CandleItem]
    has_more: bool
    venue: str
    market: str
    # The catalog instrument definition's own decimals (never derived from a value): every chart
    # label formats at these (Story 32.5).
    price_precision: int
    size_precision: int


@router.get("/api/candles/{instrument_id}")
def get_candles(
    instrument_id: str,
    before_ns: int,
    limit: int = 120,
    bar_seconds: int = 60,
) -> CandlesResponse:
    limit = max(1, min(limit, _MAX_CANDLES_LIMIT))
    bar_seconds = max(1, min(bar_seconds, _MAX_BAR_SECONDS))
    venue_of(instrument_id)  # a malformed id is a 400 (the app's ValueError handler), not a 404
    # Known limit: the instrument definition is read from the catalog on every candles request, to
    # label prices at its precision. Upgrade path: a per-instrument cache invalidated when a newer
    # definition is written (the capture service's definition refresh), so a request reads memory.
    try:
        precision = instrument_precision(CATALOG_PATH, instrument_id)
    except NoInstrumentDefinition as exc:
        # No definition: no precision to label prices at, and none is guessed (DATA-01).
        raise HTTPException(status_code=404, detail=str(exc)) from exc
    try:
        kept, has_more = chart_series.candle_page(
            instrument_id,
            before_ns,
            limit,
            bar_seconds,
            catalog_path=CATALOG_PATH,
            candles_dir=CANDLES_DB_DIR,
            recent_rows=buses.live_candle_bus.recent_rows,
            recent_liquidations=buses.live_candle_bus.recent_liquidations,
        )
    except chart_series.ImpossibleCandle as exc:
        # An impossible candle means upstream code malfunctioned (DATA-07): views has already
        # ledgered it; fail the request so the chart shows an error, never serve or drop it.
        raise HTTPException(status_code=500, detail=str(exc)) from exc
    except catalog_files.CatalogReadError as exc:
        # An archive fault, not a client error: ledgered here, where the request fails (DATA-07).
        error_ledger.record("data_api.candles_catalog_read", str(exc), exc)
        raise HTTPException(status_code=500, detail=str(exc)) from exc
    if not kept:
        return CandlesResponse(
            items=[],
            has_more=False,
            venue=venue_of(instrument_id),
            market=market_kind(instrument_id),
            price_precision=precision.price_precision,
            size_precision=precision.size_precision,
        )
    return CandlesResponse(
        items=[CandleItem(**row) for row in chart_series.with_gap_markers(kept, bar_seconds)],
        has_more=has_more,
        venue=venue_of(instrument_id),
        market=market_kind(instrument_id),
        price_precision=precision.price_precision,
        size_precision=precision.size_precision,
    )